chore: Rename ContactForms.css and ContactHero.css to follow naming convention
