+++ CP1/src/components/contact/ContactForms.css
@@ -1,27 +1,13 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap');
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: 0;
-}
-html{
-    font-size: 62.5%;
-}
 
 :root{
     --cor-btn: #6975E8;
     --cor-btn-hover: #3845c0;
 }
 
-body{
-    font-family: "Roboto", sans-serif;
-    font-weight: 400;
-    font-style: normal;
-}
+
 h1,h2,h3{
-    font-family: "Roboto", sans-serif;
     font-weight: 700;
-    font-style: normal;
 }
