+++ CP1/src/components/contact/ContactHero.css
@@ -1,35 +1,18 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap');
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: 0;
-}
-html{
-    font-size: 62.5%;
-}
+
 
 :root{
     --cor-btn: #6975E8;
 }
 
-body{
-    font-family: "Roboto", sans-serif;
-    font-weight: 400;
-    font-style: normal;
-}
-h1,h2,h3{
-    font-family: "Roboto", sans-serif;
-    font-weight: 700;
-    font-style: normal;
-}
 
 h1{
-    font-size: 7em;
+    font-size: 7rem;
     color: #333;
     line-height: 8rem;
 }
 
-img{
+.imgs{
     width: 9%;
     cursor: pointer;
 }
@@ -40,7 +23,7 @@ img{
     gap: 4rem; 
 }
 
-.hero{
+.hero1{
     display: flex;
     flex-direction: column;
     gap: 2rem;
