+++ CP1/src/index.css
@@ -73,7 +73,10 @@ header p:hover {
 .text p {
   grid-area: texto;
 }
-
+img {
+  display: flex;
+  max-width: 100%;
+}
 .text a {
   grid-area: but;
   text-decoration: none;
